+++ includes/frontend/template-tags.php
@@ -41,6 +41,12 @@ function get_the_last_modified_info () {
         $replace_ago_tt = ' ago';
     }
 
+    if( isset( $options['lmt_enable_tt_author_hyperlink_target'] ) && $options['lmt_enable_tt_author_hyperlink_target'] == 'self' ) {
+        $target = '_self';
+    } else {
+        $target = '_blank';
+    }
+
     $lmt_tt_uca = '';
 
     if( isset($options['lmt_show_author_tt_cb']) && ($options['lmt_show_author_tt_cb'] == 'default' ) ) {
@@ -48,28 +54,46 @@ function get_the_last_modified_info () {
         $author_id = get_post_meta(get_the_ID(), '_edit_last', true);
 
         if( isset($options['lmt_enable_tt_author_hyperlink']) && ($options['lmt_enable_tt_author_hyperlink'] == 'author_page' ) ) {
-            $lmt_tt_uca = $author_sep_tt . ' <a href="' . get_author_posts_url( $author_id ) . '" target="_blank" rel="author">' . get_the_author_meta( 'display_name', $author_id ) . '</a>';
+            
+            $lmt_tt_uca = $author_sep_tt . ' <a href="' . get_author_posts_url( $author_id ) . '" target="' . $target . '" rel="author">' . get_the_author_meta( 'display_name', $author_id ) . '</a>';
+        
         } elseif( isset($options['lmt_enable_tt_author_hyperlink']) && ($options['lmt_enable_tt_author_hyperlink'] == 'author_website' ) ) {
-            $lmt_tt_uca = $author_sep_tt . ' <a href="' . get_the_author_meta( 'url', $author_id ) . '" rel="author">' . get_the_author_meta( 'display_name', $author_id ) . '</a>';
+           
+            $lmt_tt_uca = $author_sep_tt . ' <a href="' . get_the_author_meta( 'url', $author_id ) . '" target="' . $target . '" rel="author">' . get_the_author_meta( 'display_name', $author_id ) . '</a>';
+        
         } elseif( isset($options['lmt_enable_tt_author_hyperlink']) && ($options['lmt_enable_tt_author_hyperlink'] == 'author_email' ) ) {
+        
             $lmt_tt_uca = $author_sep_tt . ' <a href="mailto:' . get_the_author_meta( 'user_email', $author_id ) . '" rel="author">' . get_the_author_meta( 'display_name', $author_id ) . '</a>';
+        
         } elseif( isset($options['lmt_enable_tt_author_hyperlink']) && ($options['lmt_enable_tt_author_hyperlink'] == 'none' ) ) {
+        
             $lmt_tt_uca = $author_sep_tt . ' ' . get_the_author_meta( 'display_name', $author_id );
+        
         }
     
     } elseif( isset($options['lmt_show_author_tt_cb']) && ($options['lmt_show_author_tt_cb'] == 'custom' ) ) {
         
         $get_author = $options['lmt_show_author_list_tt'];
         
+        
         if( isset($options['lmt_enable_tt_author_hyperlink']) && ($options['lmt_enable_tt_author_hyperlink'] == 'author_page' ) ) {
-            $lmt_tt_uca = $author_sep_tt . ' <a href="' . get_author_posts_url( $get_author ) . '" target="_blank" rel="author">' . get_the_author_meta( 'display_name', $get_author ) . '</a>';
+        
+            $lmt_tt_uca = $author_sep_tt . ' <a href="' . get_author_posts_url( $get_author ) . '" target="' . $target . '" rel="author">' . get_the_author_meta( 'display_name', $get_author ) . '</a>';
+        
         } elseif( isset($options['lmt_enable_tt_author_hyperlink']) && ($options['lmt_enable_tt_author_hyperlink'] == 'author_website' ) ) {
-            $lmt_tt_uca = $author_sep_tt . ' <a href="' . get_the_author_meta( 'url', $get_author ) . '" rel="author">' . get_the_author_meta( 'display_name', $get_author ) . '</a>';
+        
+            $lmt_tt_uca = $author_sep_tt . ' <a href="' . get_the_author_meta( 'url', $get_author ) . '" target="' . $target . '" rel="author">' . get_the_author_meta( 'display_name', $get_author ) . '</a>';
+        
         } elseif( isset($options['lmt_enable_tt_author_hyperlink']) && ($options['lmt_enable_tt_author_hyperlink'] == 'author_email' ) ) {
+        
             $lmt_tt_uca = $author_sep_tt . ' <a href="mailto:' . get_the_author_meta('user_email', $get_author) . '" rel="author">' . get_the_author_meta( 'display_name', $get_author ) . '</a>';
+        
         } elseif( isset($options['lmt_enable_tt_author_hyperlink']) && ($options['lmt_enable_tt_author_hyperlink'] == 'none' ) ) {
+        
             $lmt_tt_uca = $author_sep_tt . ' ' . get_the_author_meta( 'display_name', $get_author );
+        
         }
+        
     }
 
     if( isset($options['lmt_tt_enable_schema_cb']) && ($options['lmt_tt_enable_schema_cb'] == 1 ) ) {
@@ -85,11 +109,11 @@ function get_the_last_modified_info () {
     }
                    
     if(!empty($options['lmt_tt_class_box'])) {
-
-        $get_custom_class = esc_html($options['lmt_tt_class_box']);              
-        $lmt_template_tag = '<span class="'. $get_custom_class .'">' . $last_modified_text . ' <time' . $schema_tt . '>' . $lmt_tt_ud . '</time>' . $lmt_tt_uca . '</span>';
+             
+        $lmt_template_tag = '<span class="'. esc_html($options['lmt_tt_class_box']) .'">' . $last_modified_text . ' <time' . $schema_tt . '>' . $lmt_tt_ud . '</time>' . $lmt_tt_uca . '</span>';
 
     } else {
+        
         $lmt_template_tag = '<span>' . $last_modified_text . ' <time' . $schema_tt . '>' . $lmt_tt_ud . '</time>' . $lmt_tt_uca . '</span>';
     }